Tested CHS simulator
returns wavefunction and density matrix.
