+++ referenceqvm/qvm_stabilizer.py
@@ -15,12 +15,9 @@
 #    limitations under the License.
 ##############################################################################
 """
-Pure QVM that only executes pyQuil programs containing Clifford group generators,
-and return the wavefunction or stabilizer.  This is based off of the paper by
-Aaronson and Gottesman Phys. Rev. A 70, 052328
+Pure QVM that only executes pyQuil programs containing Clifford group generators.
+This is based off of the paper by Aaronson and Gottesman Phys. Rev. A 70, 052328
 """
-import sys
-import numpy as np
 from functools import reduce
 from pyquil.quil import Program, get_classical_addresses_from_program
 from pyquil.quilbase import *
@@ -30,7 +27,9 @@ from pyquil.wavefunction import Wavefunction
 from referenceqvm.qam import QAM
 from referenceqvm.gates import stabilizer_gate_matrix
 from referenceqvm.unitary_generator import value_get, tensor_up
-from referenceqvm.state_actions import project_stabilized_state
+from referenceqvm.stabilizer_utils import (project_stabilized_state,
+                                           symplectic_inner_product,
+                                           binary_stabilizer_to_pauli_stabilizer)
 
 
 class QVM_Stabilizer(QAM):
@@ -79,7 +78,7 @@ class QVM_Stabilizer(QAM):
         unitary check QVM checks makes it annoying to allow measurements.
         TODO: Change load_program for each qvm-subclass
 
-        :param Program pyquil_program: program to be ran
+        :param Program pyquil_program: program to be run
         """
         # typecheck
         if not isinstance(pyquil_program, Program):
@@ -127,8 +126,10 @@ class QVM_Stabilizer(QAM):
         """
         Construct an empty tableau for a n-qubit system
 
-        :param num_qubits:
-        :return:
+        :param Int num_qubits: number of qubits represented by the tableau
+        :return: a numpy integer array representing the tableau stabilizing
+                 the |000..000> state
+        :rtype: np.ndarray
         """
         tableau = np.zeros((2 * num_qubits, (2 * num_qubits) + 1),
                                 dtype=int)
@@ -143,7 +144,6 @@ class QVM_Stabilizer(QAM):
         Implements a full transition, including pre/post noise hooks.
 
         :param QuilAction instruction: instruction to be executed
-
         :return: if program is halted TRUE, else FALSE
         :rtype: bool int
         """
@@ -203,15 +203,17 @@ class QVM_Stabilizer(QAM):
 
     def run(self, pyquil_program, classical_addresses=None, trials=1):
         """
-        Run program.
+        Run pyquil program and return the results
 
         Loads and checks program if all gates are within the stabilizer set.
-        Then executes program
+        Then executes program.  If measurements are requested then the measured
+        results are returned
 
-        :param porgram:
-        :param classical_addresses:
-        :param trials:
-        :return:
+        :param Program pyquil_program: a pyquil Program containing only
+                                       CNOT-H-S-MEASUREMENT operations
+        :param classical_addresses: classical addresses to return
+        :param trials: number of times to repeat the execution of the program
+        :return: list of lists of classical memory after each run
         """
         self.load_program(pyquil_program)
 
@@ -236,9 +238,11 @@ class QVM_Stabilizer(QAM):
         Run program and compute the density matrix
 
         Loads and checks program if all gates are within the stabilizer set.
-        Then executes program
+        Then executes program and returns the final density matrix for the
+        stabilizer
 
-        :param porgram:
+        :param Program pyquil_program: a pyquil Program containing only
+                                       CNOT-H-S-MEASUREMENT operations
         :return:
         """
         self.load_program(pyquil_program)
@@ -251,7 +255,7 @@ class QVM_Stabilizer(QAM):
 
     def wavefunction(self, program):
         """
-        Simulate the stabilizer program and then project out the final state.
+        Simulate the program and then project out the final state.
 
         Return the final state as a wavefunction pyquil object
 
@@ -288,7 +292,7 @@ class QVM_Stabilizer(QAM):
 
         :param Int h: row index 'h'
         :param Int i: row index 'i'
-        :return:
+        :return: None but mutates the tableau
         """
         # NOTE: this is left multiplication of P(i).  P(i) * P(h)
         phase_accumulator = self._rowsum_phase_accumulator(h, i)
@@ -311,6 +315,7 @@ class QVM_Stabilizer(QAM):
         phase accumulator sub algorithm for the rowsum routine
 
         Note: this accumulates the $i$ phase for row_i * row_h  NOT row_h, row_i
+
         :param Int h: row index 'h'
         :param Int i: row index 'i'
         :return: phase mod 4
@@ -332,11 +337,11 @@ class QVM_Stabilizer(QAM):
 
         when the pauli matrices represented by x1z1 and x2z2 are multiplied together
 
-        :param x1:
-        :param z1:
-        :param x2:
-        :param z2:
-        :return:
+        :param x1: binary variable for the first pauli operator representing X
+        :param z1: binary variable for th first pauli operator representing Z
+        :param x2: binary variable for the second pauli operator representing X
+        :param z2: binary variable for the second pauli operator representing Z
+        :return: power that 1j is raised to when multiplying paulis together
         """
         # if the first term is identity
         if x1 == z1 == 0:
@@ -366,13 +371,14 @@ class QVM_Stabilizer(QAM):
         if x1 == 0 and z1 == 1:
             return x2 * (1 - 2 * z2)
 
-        raise ValueError("we were unable to multiply the pauli operators together!")
+        raise ValueError("We were unable to multiply the pauli operators" +
+                         " together!")
 
     def _apply_cnot(self, instruction):
         """
         Apply a CNOT to the tableau
 
-        :param instruction: pyquil abstract instruction.  Must have
+        :param instruction: pyquil abstract instruction.
         """
         a, b = [value_get(x) for x in instruction.qubits]
         for i in range(2 * self.num_qubits):
@@ -382,7 +388,7 @@ class QVM_Stabilizer(QAM):
 
     def _cnot_phase_update(self, i, c, t):
         """
-        update r_{i}
+        update r_{i} as a submethod to applying CNOT to the tableau
 
         :param i: tableau row index
         :param c: control qubit index
@@ -396,8 +402,7 @@ class QVM_Stabilizer(QAM):
         """
         Apply a hadamard gate on qubit defined in instruction
 
-        :param instruction:
-        :return:
+        :param instruction: pyquil abstract instruction.
         """
         qubit_label = [value_get(x) for x in instruction.qubits][0]
         for i in range(2 * self.num_qubits):
@@ -408,8 +413,7 @@ class QVM_Stabilizer(QAM):
         """
         Apply the phase gate instruction ot the tableau
 
-        :param instruction:
-        :return:
+        :param instruction: pyquil abstract instruction.
         """
         qubit_label = [value_get(x) for x in instruction.qubits][0]
         for i in range(2 * self.num_qubits):
@@ -417,6 +421,11 @@ class QVM_Stabilizer(QAM):
             self.tableau[i, qubit_label + self.num_qubits] = self.tableau[i, qubit_label + self.num_qubits] ^ self.tableau[i, qubit_label]
 
     def _apply_measurement(self, instruction):
+        """
+        Apply a measurement
+
+        :param instruction: pyquil abstract instruction.
+        """
         t_qbit = value_get(instruction.qubit)
         t_cbit = value_get(instruction.classical_reg)
 
@@ -424,9 +433,9 @@ class QVM_Stabilizer(QAM):
         # this is analogous to the case when the measurement operator does not
         # commute with at least one stabilizer
         if any(self.tableau[self.num_qubits:, t_qbit] == 1):
-            # find the first one.
+            # find the first `1'.
             xpa_idx = np.where(self.tableau[self.num_qubits:, t_qbit] == 1)[0][0]  # take the first index
-            xpa_idx += self.num_qubits  # adjust so we can index into the tableau
+            xpa_idx += self.num_qubits   # adjust so we can index into the tableau
             for ii in range(2 * self.num_qubits):  # loop over each row and call rowsum(ii, xpa_idx)
                 if ii != xpa_idx and self.tableau[ii, t_qbit] == 1:
                     self._rowsum(ii, xpa_idx)
@@ -447,223 +456,24 @@ class QVM_Stabilizer(QAM):
 
         # outcome of measurement is deterministic...need to determine sign
         else:
-            # augment tableaue with a scratch space
+            # augment tableau with a scratch space
             self.tableau = np.vstack((self.tableau, np.zeros((1, 2 * self.num_qubits + 1), dtype=int)))
             for ii in range(self.num_qubits):
                 # We need to check if R(i) anticommutes with Za...which it does if x_{ia} = 1
-                if self.tableau[ii, t_qbit] == 1:  # refrencing the destabilizers
+                if self.tableau[ii, t_qbit] == 1:  # referencing the destabilizers
 
+                    # TODO: Remove this for performance?
                     # check something silly.  Does the destabilizer anticommute with the observable?  It SHOULD!
                     tmp_vector_representing_z_qubit = np.zeros((2 * self.num_qubits), dtype=int)
                     tmp_vector_representing_z_qubit[t_qbit] = 1
                     assert symplectic_inner_product(tmp_vector_representing_z_qubit, self.tableau[ii, :-1]) == 1
 
                     # row sum on the stabilizers (summing up operators such that we get operator Z_{a})
-                    self._rowsum(2 * self.num_qubits, ii + self.num_qubits)  # note: A-G says 2 n + 1...this is correct...but they start counting at 1 not zero
+                    # note: A-G says 2 n + 1...this is correct...but they start counting at 1 not zero
+                    self._rowsum(2 * self.num_qubits, ii + self.num_qubits)
 
             # set the classical bit to be the last element of the scratch row
             self.classical_memory[t_cbit] = self.tableau[-1, -1]
 
             # remove the scratch space
             self.tableau = self.tableau[:2 * self.num_qubits, :]
-
-
-
-def pauli_stabilizer_to_binary_stabilizer(stabilizer_list):
-    """
-    Convert a list of stabilizers represented as PauliTerms to a binary tableau form
-
-    :param List stabilizer_list: list of stabilizers where each element is a PauliTerm
-    :return: return an integer matrix representing the stabilizers where each row is a
-             stabilizer.  The size of the matrix is n x (2 * n) where n is the maximum
-             qubit index.
-    """
-    if not all([isinstance(x, PauliTerm) for x in stabilizer_list]):
-        raise TypeError("At least one element of stabilizer_list is not a PauliTerm")
-
-    max_qubit = max([max(x.get_qubits()) for x in stabilizer_list]) + 1
-    stabilizer_tableau = np.zeros((len(stabilizer_list), 2 * max_qubit + 1), dtype=int)
-    for row_idx, term in enumerate(stabilizer_list):
-        for i, pterm in term:  # iterator for each tensor-product element of the Pauli operator
-            if pterm == 'X':
-                stabilizer_tableau[row_idx, i] = 1
-            elif pterm == 'Z':
-                stabilizer_tableau[row_idx, i + max_qubit] = 1
-            elif pterm == 'Y':
-                stabilizer_tableau[row_idx, i] = 1
-                stabilizer_tableau[row_idx, i + max_qubit] = 1
-            else:
-                # term is identity
-                pass
-
-        if not (np.isclose(term.coefficient, -1) or np.isclose(term.coefficient, 1)):
-            raise ValueError("stabilizers must have a +/- coefficient")
-
-        if int(np.sign(term.coefficient.real)) == 1:
-            stabilizer_tableau[row_idx, -1] = 0
-        elif int(np.sign(term.coefficient.real)) == -1:
-            stabilizer_tableau[row_idx, -1] = 1
-        else:
-            raise TypeError('unrecognized on pauli term of stabilizer')
-
-    return stabilizer_tableau
-
-
-def binary_stabilizer_to_pauli_stabilizer(stabilizer_tableau):
-    """
-    Convert a stabilizer tableau to a list of PauliTerms
-
-    :param stabilizer_tableau:
-    :return:
-    """
-    stabilizer_list = []
-    num_qubits = (stabilizer_tableau.shape[1] - 1) // 2
-    for nn in range(stabilizer_tableau.shape[0]):  # iterate through the rows
-        stabilizer_element = []
-        for ii in range(num_qubits):
-            if stabilizer_tableau[nn, ii] == 1 and stabilizer_tableau[nn, ii + num_qubits] == 0:
-                stabilizer_element.append(sX(ii))
-            elif stabilizer_tableau[nn, ii] == 0 and stabilizer_tableau[nn, ii + num_qubits] == 1:
-                stabilizer_element.append(sZ(ii))
-            elif stabilizer_tableau[nn, ii] == 1 and stabilizer_tableau[nn, ii + num_qubits] == 1:
-                stabilizer_element.append(sY(ii))
-
-        stabilizer_term = reduce(lambda x, y: x * y, stabilizer_element) * ((-1) ** stabilizer_tableau[nn, -1])
-        stabilizer_list.append(stabilizer_term)
-    return stabilizer_list
-
-
-def binary_rref(code_matrix):
-    """
-    Convert the binary code_matrix into rref
-
-    :param code_matrix:
-    :return:
-    """
-    code_matrix = code_matrix.astype(int)
-    rdim, cdim = code_matrix.shape
-    for ridx in range(rdim):
-        # print("start [{},{}] = {}".format(ridx, ridx,
-        #                                   code_matrix[ridx, ridx]))
-        # set first element
-        if not np.isclose(code_matrix[ridx, ridx], 1):
-            for ii in range(ridx + 1, rdim):
-                if np.isclose(code_matrix[ii, ridx], 1.0):
-                    # switch rows
-                    # code_matrix[ridx, :], code_matrix[ii, :] = code_matrix[ii, :], code_matrix[ridx, :]
-                    code_matrix[[ridx, ii]] = code_matrix[[ii, ridx]]
-                    print("r_{} <-> r_{}".format(ii + 1, ridx + 1))
-
-        # print(code_matrix)
-        # start elimination of all other rows
-        # print("pivot [{}, {}] = {}".format(ridx, ridx,
-        #                                    code_matrix[ridx, ridx]))
-
-        rows_to_eliminate = set(range(rdim))
-        rows_to_eliminate.remove(ridx)
-        # print(rows_to_eliminate)
-        for elim_ridx in list(rows_to_eliminate):
-            # eliminate by Hadamard product on the rows (XOR element wise)
-            # if other row element is 1 then eliminate by subtracting ridx row
-
-            # if np.isclose(code_matrix[elim_ridx, ridx], 1.0):
-            #     code_matrix[elim_ridx, :] = code_matrix[elim_ridx, :] - \
-            #                                 code_matrix[ridx, :]
-            if np.isclose(code_matrix[elim_ridx, ridx], 1.0):
-                code_matrix[elim_ridx, :] = code_matrix[elim_ridx, :] ^ code_matrix[ridx, :]
-                # print("elim [{}, :]".format(elim_ridx))
-
-        # print("semi rref at row {}".format(ridx))
-        # print(code_matrix)
-    return code_matrix
-
-
-def symplectic_inner_product(vector1, vector2):
-    """
-    Operators commute if their binary form symplectic inner product is zero
-
-    Operators anticommute if their binary form symplectic inner product is one
-
-    :param vector1: binary form of operator with no sign info
-    :param vector2: binary form of a pauli operator with no sign info
-    :return: 0, 1
-    """
-    if vector1.shape != vector2.shape:
-        raise ValueError("vectors must be the same size.")
-
-    # TODO: add a check for binary or integer linear arrays
-
-    hadamard_product = np.multiply(vector1, vector2)
-    return reduce(lambda x, y: x ^ y, hadamard_product)
-
-
-if __name__ == "__main__":
-    # practice my clifford transformations
-    # <Z1,Z2,...Zn> = |0000...0N>
-    # N |psi> = N g |psi> = NgNd N|psi> = g2 |psi2> = |psi2>
-
-    from referenceqvm.gates import X, Y, Z, H, CNOT, S, I
-    from sympy import Matrix
-
-    # HZH = X
-    test_X = H.dot(Z).dot(H)
-    assert np.allclose(test_X, X)
-
-    # HXH = Z
-    test_Z = H.dot(X).dot(H)
-    assert np.allclose(test_Z, Z)
-
-    # HYH = iHZHHXH = iXZ = -Y
-    test_nY = H.dot(Y).dot(H)
-    assert np.allclose(test_nY, -Y)
-
-    # XYX = -Y
-    assert np.allclose(X.dot(Y).dot(X), -Y)
-
-    # XZX = -Z
-    assert np.allclose(X.dot(Z).dot(X), -Z)
-
-    # SXS = Y
-    assert np.allclose(S.dot(X).dot(np.conj(S).T), Y)
-
-    # CNOT(X otimes X)CNOT = X otimes I
-    # (C, T) for CNOT as written in reference-qvm
-    assert np.allclose(CNOT.dot(np.kron(I, X)).dot(CNOT), np.kron(I, X))
-    assert np.allclose(CNOT.dot(np.kron(X, I)).dot(CNOT), np.kron(X, X))
-    assert np.allclose(CNOT.dot(np.kron(X, X)).dot(CNOT), np.kron(X, I))
-
-    # CNOT(Z otimes Z)CNOT = Z otimes I
-    assert np.allclose(CNOT.dot(np.kron(Z, Z)).dot(CNOT), np.kron(I, Z))
-    assert np.allclose(CNOT.dot(np.kron(Z, I)).dot(CNOT), np.kron(Z, I))
-    assert np.allclose(CNOT.dot(np.kron(I, Z)).dot(CNOT), np.kron(Z, Z))
-
-    Z1 = np.kron(Z, I)
-    Z2 = np.kron(I, Z)
-    ZZ = np.kron(Z, Z)
-    X1 = np.kron(X, I)
-    X2 = np.kron(I, X)
-    XX = np.kron(X, X)
-    YY = np.kron(Y, Y)
-
-    generator_matrix = np.zeros((3, 4))
-    generator_matrix[0, 2] = generator_matrix[0, 3] = 1  # ZZ
-    generator_matrix[1, 0] = generator_matrix[1, 1] = 1  # XX
-    generator_matrix[2, 0] = generator_matrix[2, 2] = 1  # Y
-    generator_matrix[2, 1] = generator_matrix[2, 3] = 1  # Y
-    parity_vector = np.zeros((3, 1))
-    parity_vector[-1] = 1
-
-    # define the amplitudes from the generator set by diagonalizing
-    # the tableau matrix?
-    generator_matrix = np.hstack((generator_matrix, parity_vector))
-    print(generator_matrix)
-
-    generator_matrix_2 = binary_rref(generator_matrix)
-    print('\n')
-    print(generator_matrix_2)
-
-    mat = Matrix(generator_matrix)
-    mat, pivots = mat.rref()
-    mat = np.asarray(mat).astype(int)
-    print(mat)
-
